Version 1.1.0

+++ plugin.video.vstream/resources/lib/tmdb.py
@@ -607,9 +607,9 @@ class cTMDb:
         try:
             duration = 0
             if 'runtime' in meta and meta['runtime']:
-                duration = int(meta['runtime'])
+                duration = float(meta['runtime'])
             elif 'episode_run_time' in meta and meta['episode_run_time']:
-                duration = int(meta['episode_run_time'][0])
+                duration = float(meta['episode_run_time'][0])
             
             if duration < 300 : # en minutes
                 duration *= 60  # Convertir les minutes TMDB en secondes pour KODI
